[docs] Update old links and soften some wording
Soften the phrase "very much outside" to just "outside" to make the tone
of the docs a little friendlier.

Switch the documentation link for `lint` from a third-party site that
doesn't have valid HTTPS certificates to a developer.android.com page.

Update the documentation links for `aapt2` to remove a broken link and
instead point to the dedicated page for `aapt2` that is now available on
developer.android.com.

Update the link about maximum path length on Windows to the current
location.