[docs] Reformat error and warning docs to match NuGet error docs
Reformat the documents to match the [NuGet error reference][0] pages.
Each of those pages starts with a code block (a pre-formatted section)
that shows the error message, followed by an `Issue` section, and then a
`Solution` section.  This format is nice because it provides a
consistent way to highlight the key information for each error code.

This commit attempts to focus on structural rearrangement, but it does
include a few wording changes for cases where the original wording
sounded out of place after the structural rearrangement.

[0]: https://docs.microsoft.com/nuget/reference/errors-and-warnings/